Generic normalization procedure

This is a generic normalization procedure with which you can create your own normalization method.

Usage

norm_generic(
  .data,
  ...,
  .by = NULL,
  .norm_fun = identity_norm_fun,
  .by_formant = FALSE,
  .drop_orig = FALSE,
  .keep_params = FALSE,
  .names = "{.col}_n",
  .silent = FALSE,
  .call = caller_env()
)

Arguments

.data

A data frame containing vowel formant data

...

<tidy-select> One or more unquoted expressions separated by commas. These should target the vowel formant data columns.

.by

<tidy-select> A selection of columns to group by. Typically a column of speaker IDs.

.norm_fun

A function to apply to a pivoted long data frame. See Details for how to write such a function.

.by_formant

Whether or not the normalization method is formant intrinsic.

.drop_orig

Whether or not to drop the original formant data columns.

.keep_params

Whether or not to keep the Location (*_L) and scale (*_S) normalization parameters

.names

A glue::glue() expression for naming the normalized data columns. The "{.col}" portion corresponds to the name of the original formant columns.

.silent

Whether or not the informational message should be printed.

.call

Used for internal purposes.

Details

The following norm_* procedures are built on top of norm_generic().

norm_lobanov

norm_nearey

norm_deltaF

You can define a custom normalization function to pass to .norm_fun that takes the following arguments

.data

.grouping

.names

The .data data frame will be your original data frame with the formant columns pivoted longer. The formant names will be available in a column called .formant, and the formant values in a column called .col. Any Location parameters should be assigned to a column called L, and any Scale parameters should be assigned to a column called S. A median-based normalization function might look like the following.

median_norm_fun <- function(.data, .grouping, .names){
  .data |>
    mutate(
      .by = !!.grouping,
      L = median(.col),
      S = mad(.col),
      across(
        .col,
        .fns = \(x) (x - L)/S,
        .names = .names
      )
    )
 return(.data)
}

If the method is meant to be formant ex/intrinsic, as defined by .by_formant, this will be accounted for by the grouping variables in the .grouping argument passed to the normalization function.

Examples

library(tidynorm)
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

median_norm_fun <- function(.data, .grouping, .names){
  .data <- .data |>
    mutate(
      .by = !!.grouping,
      L = median(.col, na.rm = TRUE),
      S = mad(.col, na.rm = TRUE),
      across(
        .col,
        .fns = \(x) (x - L)/S,
        .names = .names
      )
    )
  return(.data)
}

speaker_data |>
  norm_generic(
    F1:F3,
    .by = speaker,
    .norm_fun = median_norm_fun,
    .by_formant = TRUE,
    .drop_orig = TRUE,
    .names = "{.col}_mad"
  )
#> Normalization info
#> • normalized `F1`, `F2`, and `F3`
#> • normalized values in `F1_mad`, `F2_mad`, and `F3_mad`
#> • grouped by `speaker`
#> • formant intrinsic
#> 
#> # A tibble: 10,697 × 9
#>      .id speaker vowel plt_vclass ipa_vclass word      F1_mad  F2_mad F3_mad
#>    <int> <chr>   <chr> <chr>      <chr>      <chr>      <dbl>   <dbl>  <dbl>
#>  1     1 s01     EY    eyF        ejF        OKAY      1.13    0.400  -0.521
#>  2     2 s01     AH    uh         ʌ          UM        0.643  -0.166   0.629
#>  3     3 s01     AY    ay         aj         I'M       2.09   -0.0211  0.165
#>  4     4 s01     IH    i          ɪ          LIVED    -0.454  -1.13    1.41 
#>  5     5 s01     IH    i          ɪ          IN       -0.0221  1.04    1.03 
#>  6     6 s01     AH    @          ə          COLUMBUS -0.0206 -0.104   0.715
#>  7     7 s01     AY    ay         aj         MY       -0.280   0.0479  0.968
#>  8     8 s01     IH    i          ɪ          ENTIRE   -0.649   1.07    1.03 
#>  9     9 s01     ER    *hr        ə˞         ENTIRE   -0.585  -0.709  -2.35 
#> 10    10 s01     AY    ay0        aj0        LIFE      0.987  -0.656  -2.34 
#> # ℹ 10,687 more rows